refactor: 모달 이름 변경
