+++ components/members/detail/MessageSection/index.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import useModalState from '@/components/common/Modal/useModalState';
 import useToast from '@/components/common/Toast/useToast';
 import MemberDetailSection from '@/components/members/detail/MemberDetailSection';
-import CoffeeChatModal from '@/components/members/detail/MessageSection/CoffeeChatModal';
+import CoffeeChatModal from '@/components/members/detail/MessageSection/MessageModal';
 import { colors } from '@/styles/colors';
 import { MOBILE_MEDIA_QUERY } from '@/styles/mediaQuery';
 import { textStyles } from '@/styles/typography';
